authentication and authorisation using jwt

+++ controllers/api/v1/posts_api.js
@@ -23,7 +23,7 @@ module.exports.index=async function(req,res){
 module.exports.destroy= async function(req,res){
     try{
         let post=await Post.findById(req.params.id);
-        // if(post.user==req.user.id){
+        if(post.user==req.user.id){
            post.remove();
            await Comment.deleteMany({post:req.params.id});
         //    if(req.xhr){
@@ -37,8 +37,10 @@ module.exports.destroy= async function(req,res){
            return res.status(200).json({
             message:"post deleted"
            });
-        // } else
-        //    return res.redirect('back');
+        } else return res.status(401).json({
+            message: "you cannot delete this post"
+        })
+          
     }catch(err){
         console.log('Error ',err);
         return res.status(200).json({
